[Tests] Align test names with implementation (Fixes #160)
